+++ src/furet/app/utils.py
@@ -3,8 +3,7 @@ from typing import Any
 from PySide6 import QtWidgets, QtCore, QtGui
 
 from furet.app.widgets.checkableComboBox import CheckableComboBox
-from furet.app.widgets.objectTableModel import ObjectTableColumn
-from furet.app.widgets.selectAllTextComboBox import SelectAllTextComboBox
+from furet.app.widgets.objectTableWidget import ObjectTableColumn
 
 from typing import TypeVar, Generic
 
@@ -14,29 +13,6 @@ from furet.types.decree import Decree
 from furet.types.department import Department
 
 T = TypeVar('T')
-
-# TODO turn into a custom widget
-def buildComboBox(options: list[T], choice: T | None, none: tuple[str, Any] | None = None) -> QtWidgets.QComboBox:
-    box = SelectAllTextComboBox()
-    if none is not None:
-        box.addItem(none[0], none[1])
-        if choice is None: 
-            box.setCurrentIndex(0)
-    for o in options:
-        box.addItem(str(o), o)
-    
-    if choice is not None:
-        for i, o in enumerate(options):
-            if o == choice:
-                box.setCurrentIndex(i if none is None else (i+1))
-
-    completer = QtWidgets.QCompleter([str(o) for o in options], box)
-    completer.setFilterMode(QtCore.Qt.MatchFlag.MatchContains)
-    completer.setCaseSensitivity(QtCore.Qt.CaseSensitivity.CaseInsensitive)
-    completer.setCompletionMode(QtWidgets.QCompleter.CompletionMode.UnfilteredPopupCompletion)
-    box.setCompleter(completer)
-
-    return box
 
 def buildMultiComboBox(options: list[T], choices: list[T], none: str = None) -> CheckableComboBox:
     box = CheckableComboBox()
