+++ src/furet/app/widgets/elidedLabel.py
@@ -1,29 +1,26 @@
 from PySide6 import QtWidgets, QtCore, QtGui
 
 class ElidedLabel(QtWidgets.QLabel):
-    _txt: str = ""
+    _text: str = ""
     _elideMode: QtCore.Qt.TextElideMode = QtCore.Qt.TextElideMode.ElideRight
 
-    def __init__(self, text: str, elideMode: QtCore.Qt.TextElideMode = QtCore.Qt.TextElideMode.ElideRight, parent: QtWidgets.QWidget = None):
-        if parent is None:
-            super().__init__("")
-        else:
-            super().__init__(parent, "")
-        self._txt = text
+    def __init__(self, text: str, /, parent: QtWidgets.QWidget | None = None, elideMode: QtCore.Qt.TextElideMode = QtCore.Qt.TextElideMode.ElideRight):
+        super().__init__("", parent)
+        self._text = text
         self._elideMode = elideMode
         self.updateDisplayedText()
 
-    def setText(self, arg__1):
-        self._txt = arg__1
+    def setText(self, arg__1, /):
+        self._text = arg__1
         self.updateDisplayedText()
 
-    def text(self):
-        return self._txt
+    def text(self, /):
+        return self._text
 
     def setElideMode(self, elidedMode: QtCore.Qt.TextElideMode):
         self._elideMode = elidedMode
 
-    def elideMode(self):
+    def elideMode(self) -> QtCore.Qt.TextElideMode:
         return self._elideMode
 
     def resizeEvent(self, event):
@@ -34,9 +31,9 @@ class ElidedLabel(QtWidgets.QLabel):
         fm = QtGui.QFontMetrics(self.fontMetrics())
         elidedText = fm.elidedText(
             self.text(), self._elideMode, self.width(), QtCore.Qt.TextFlag.TextShowMnemonic)
-        if (len(self._txt) > 0):
-            showFirstCharacter = self._txt[0] + "..."
+        if (len(self._text) > 0):
+            showFirstCharacter = self._text[0] + "..."
             self.setMinimumWidth(fm.horizontalAdvance(showFirstCharacter) + 1)
         if self.textInteractionFlags() == QtCore.Qt.TextInteractionFlag.TextBrowserInteraction:
-            elidedText = f'<a href="{self._txt}">{elidedText}</a>'
+            elidedText = f'<a href="{self._text}">{elidedText}</a>'
         super().setText(elidedText)
